chore(refactor): UI enhacements
