+++ src/components/Dashboards/Dashboard.tsx
@@ -10,8 +10,8 @@ import { useWidgets } from '../../context/WidgetsContext';
 import type { WidgetProps } from '../../types/types';
 import { NoFocusIconButton } from '../common/NoFocusIconButton';
 import EmptyPlaceHolder from '../EmptyPlaceHolder/EmptyPlaceHolder';
-import ImportWidgetModal from '../Modal/ImportWidgetModal';
-import ViewWidgetConfigurationModal from '../Modal/ViewWidgetConfigurationModal';
+import ImportWidgetModal from '../Modals/ImportWidgetModal';
+import ViewWidgetConfigurationModal from '../Modals/ViewWidgetConfigurationModal';
 import CardWidget from '../Widgets/CardWidget/CardWidget';
 import ChartWidget from '../Widgets/ChartWidget/ChartWidget';
 import NumberWidget from '../Widgets/NumberWidget/NumberWidget';
