+++ src/components/Modals/ViewWidgetConfigurationModal.tsx
